agt-auth/requirements.txt
# Framework
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Base de données
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache / Sessions
django-redis==5.4.0
redis==5.0.4

# JWT
PyJWT==2.8.0
cryptography==42.0.5

# Hashing
bcrypt==4.1.3

# 2FA / TOTP
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.3.0

# API Documentation (Swagger/OpenAPI 3.0)
drf-spectacular==0.27.2

# Validation & Utilities
phonenumbers==8.13.37
python-decouple==3.8

# Logging
python-json-logger==2.0.7

# HTTP client (appels inter-services)
httpx==0.27.0

# Serveur WSGI production
gunicorn==22.0.0

# Tests
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
faker==25.0.0
coverage==7.5.1
